refactor: 비밀번호 재설정 컴포넌트 리팩토링
